requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.22
numpy==1.24.3
python-dotenv==1.0.0
